+++ docker/scripts/install-connectors.sh
@@ -18,12 +18,14 @@ function cmakebuild() {
 
 cd /tmp
 
-BUILD_PACKAGES="git cmake make gcc g++ musl-dev"
+BUILD_PACKAGES="git cmake make gcc g++"
 
-apk add --no-cache --virtual .build-deps $BUILD_PACKAGES
-
+apt-get update
+apt-get -y install --no-install-recommends $BUILD_PACKAGES
 
 git clone https://github.com/jketterl/owrx_connector.git
-cmakebuild owrx_connector 9d72cf1382ed90735632a6d0ef6f820a4758f733
+cmakebuild owrx_connector 45ec227b38bb763b0a923a1856740f4ddf74216c
 
-apk del .build-deps
+apt-get -y purge --autoremove $BUILD_PACKAGES
+apt-get clean
+rm -rf /var/lib/apt/lists/*