+++ docker/scripts/install-dependencies-soapyremote.sh
@@ -18,13 +18,15 @@ function cmakebuild() {
 
 cd /tmp
 
-STATIC_PACKAGES="avahi"
-BUILD_PACKAGES="git cmake make gcc musl-dev g++ linux-headers avahi-dev"
+STATIC_PACKAGES="avahi-daemon libavahi-client3"
+BUILD_PACKAGES="git cmake make gcc g++ libavahi-client-dev"
 
-apk add --no-cache $STATIC_PACKAGES
-apk add --no-cache --virtual .build-deps $BUILD_PACKAGES
+apt-get update
+apt-get -y install --no-install-recommends $STATIC_PACKAGES $BUILD_PACKAGES
 
 git clone https://github.com/pothosware/SoapyRemote.git
 cmakebuild SoapyRemote 6d9bd820da470cfe7b27b2e6946af93cfece448f
 
-apk del .build-deps
+apt-get -y purge --autoremove $BUILD_PACKAGES
+apt-get clean
+rm -rf /var/lib/apt/lists/*