+++ docker/scripts/install-dependencies-soapysdr.sh
@@ -18,13 +18,15 @@ function cmakebuild() {
 
 cd /tmp
 
-STATIC_PACKAGES="udev"
+STATIC_PACKAGES="libudev1"
 BUILD_PACKAGES="git cmake make patch wget sudo gcc g++"
 
-apk add --no-cache $STATIC_PACKAGES
-apk add --no-cache --virtual .build-deps $BUILD_PACKAGES
+apt-get update
+apt-get -y install --no-install-recommends $STATIC_PACKAGES $BUILD_PACKAGES
 
 git clone https://github.com/pothosware/SoapySDR
 cmakebuild SoapySDR f722f9ce5b629c3c44401a9bf628b3f8e67a9695
 
-apk del .build-deps
+SUDO_FORCE_REMOVE=yes apt-get -y purge --autoremove $BUILD_PACKAGES
+apt-get clean
+rm -rf /var/lib/apt/lists/*