+++ docker/scripts/install-dependencies.sh
@@ -18,24 +18,46 @@ function cmakebuild() {
 
 cd /tmp
 
-STATIC_PACKAGES="sox fftw python3 netcat-openbsd libsndfile lapack libusb qt5-qtbase qt5-qtmultimedia qt5-qtserialport qt5-qttools alsa-lib"
-BUILD_PACKAGES="git libsndfile-dev fftw-dev cmake ca-certificates make gcc musl-dev g++ lapack-dev linux-headers autoconf automake libtool texinfo gfortran libusb-dev qt5-qtbase-dev qt5-qtmultimedia-dev qt5-qtserialport-dev qt5-qttools-dev asciidoctor asciidoc alsa-lib-dev linux-headers"
+STATIC_PACKAGES="sox libfftw3-bin python3 python3-setuptools netcat-openbsd libsndfile1 liblapack3 libusb-1.0-0 libqt5core5a libreadline7 libgfortran4 libgomp1 libasound2 libudev1 ca-certificates"
+BUILD_PACKAGES="wget git libsndfile1-dev libfftw3-dev cmake make gcc g++ liblapack-dev autoconf automake libtool texinfo gfortran libusb-1.0-0-dev qtbase5-dev qtmultimedia5-dev qttools5-dev libqt5serialport5-dev qttools5-dev-tools asciidoctor asciidoc libasound2-dev pkg-config libudev-dev libhamlib-dev patch xsltproc"
 
-apk add --no-cache $STATIC_PACKAGES
-apk add --no-cache --virtual .build-deps $BUILD_PACKAGES
+apt-get update
+apt-get -y install auto-apt-proxy
+apt-get -y install --no-install-recommends $STATIC_PACKAGES $BUILD_PACKAGES
+
+case `uname -m` in
+    arm*)
+        PLATFORM=armhf
+        ;;
+    aarch64*)
+        PLATFORM=aarch64
+        ;;
+    x86_64*)
+        PLATFORM=amd64
+        ;;
+esac
+
+pushd /tmp
+wget https://github.com/just-containers/s6-overlay/releases/download/v1.21.8.0/s6-overlay-${PLATFORM}.tar.gz
+tar xzf s6-overlay-${PLATFORM}.tar.gz -C /
+rm s6-overlay-${PLATFORM}.tar.gz
+popd
 
 git clone https://github.com/jketterl/js8py.git
 pushd js8py
 git checkout 888e62be375316882ad2b2ac8e396c3bf857b6fc
 python3 setup.py install
 popd
+rm -rf js8py
 
 git clone https://git.code.sf.net/p/itpp/git itpp
 cmakebuild itpp bb5c7e95f40e8fdb5c3f3d01a84bcbaf76f3676d
 
 git clone https://github.com/jketterl/csdr.git
 cd csdr
-git checkout fe0b042d9cdc2605a817ca7fdd3a23c48bf07563
+git checkout c4d8a8a5590898e8c9e94b88b96a2fdc7cd0493a
+autoreconf -i
+./configure
 make
 make install
 cd ..
@@ -50,21 +72,14 @@ cmakebuild digiham 95206501be89b38d0267bf6c29a6898e7c65656f
 git clone https://github.com/f4exb/dsd.git
 cmakebuild dsd f6939f9edbbc6f66261833616391a4e59cb2b3d7
 
-git clone https://github.com/Hamlib/Hamlib.git
-pushd Hamlib
-git checkout 301ebb92eaa538dfa75c06821f46715f40dd7673
-./bootstrap
-./configure
-make
-make install
-popd
-rm -rf Hamlib
-
 JS8CALL_VERSION=2.1.1
 JS8CALL_DIR=js8call-${JS8CALL_VERSION}
 JS8CALL_TGZ=${JS8CALL_DIR}.tgz
 wget http://files.js8call.com/${JS8CALL_VERSION}/${JS8CALL_TGZ}
 tar xfz ${JS8CALL_TGZ}
+# patch allows us to build against the packaged hamlib
+patch -Np1 -d ${JS8CALL_DIR} < /js8call-hamlib.patch
+rm /js8call-hamlib.patch
 CMAKE_ARGS="-D CMAKE_CXX_FLAGS=-DJS8_USE_LEGACY_HAMLIB" cmakebuild ${JS8CALL_DIR}
 rm ${JS8CALL_TGZ}
 
@@ -72,12 +87,13 @@ WSJT_DIR=wsjtx-2.1.2
 WSJT_TGZ=${WSJT_DIR}.tgz
 wget http://physics.princeton.edu/pulsar/k1jt/${WSJT_TGZ}
 tar xfz ${WSJT_TGZ}
+patch -Np0 -d ${WSJT_DIR} < /wsjtx-hamlib.patch
+mv /wsjtx.patch ${WSJT_DIR}
 cmakebuild ${WSJT_DIR}
 rm ${WSJT_TGZ}
 
 git clone --depth 1 -b 1.5 https://github.com/wb2osz/direwolf.git
 cd direwolf
-patch -Np1 < /direwolf-1.5.patch
 make
 make install
 cd ..
@@ -88,4 +104,6 @@ pushd /opt/aprs-symbols
 git checkout 5c2abe2658ee4d2563f3c73b90c6f59124839802
 popd
 
-apk del .build-deps
+apt-get -y purge --autoremove $BUILD_PACKAGES
+apt-get clean
+rm -rf /var/lib/apt/lists/*