+++ CHANGELOG.md
@@ -20,6 +20,11 @@
   [the js8py library](https://github.com/jketterl/js8py))
 - Reorganization of the frontend demodulator code
 - Improve receiver load time by concatenating javascript assets
+- Docker images migrated to Debian slim images; This was necessary to allow the use of function multiversioning in
+  csdr and owrx_connector to allow the images to run on a wider range of CPUs
+- Docker containers have been updated to include the SDRplay driver version 3
+- HackRF support is now based on SoapyHackRF
+- Removed sdr.hu server listing support since the site has been shut down
 
 **0.18.0**
 - Support for SoapyRemote